// src/api/__generated__/wallet/model/cursorPageResponseNotificationItemResponse.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Mini Drama User Wallet API
 * 用户、钱包、任务中心接口
 * OpenAPI spec version: v1.0
 */
import type { NotificationItemResponse } from './notificationItemResponse';

/**
 * 游标分页响应
 */
export interface CursorPageResponseNotificationItemResponse {
  /** 每页数量 */
  pageSize?: number;
  /** 下一页游标，-1 表示没有更多数据 */
  mark?: string;
  /** 是否还有更多数据 */
  hasMore?: boolean;
  /** 分页数据 */
  list?: NotificationItemResponse[];
}